Make store interact with recipe service
Instead of RecipeView calling recipe service functions directly make
the view call store when creating/updating/deleting recipes. The store
calls the recipe service (which connects to Firestore) and updates
the store data as needed.

Next probably need to make the recipe objects aware of their batch
(and update the batches also on the client when making changes).